requirements.txt
Django~=3.0.5
djangorestframework==3.11.0
pygments==2.5.2
psycopg2==2.7.6.1
gunicorn==19.7.1
heroku==0.1.4
